+++ docs/Changelog.md
@@ -4,6 +4,8 @@
 
 - Mapsforge: deduplicate maps [#903](https://github.com/mapsforge/vtm/pull/903)
 - Fix overlapping map regions [#903](https://github.com/mapsforge/vtm/pull/903) [#905](https://github.com/mapsforge/vtm/pull/905)
+- Mapsforge: simplification exceptions [#906](https://github.com/mapsforge/vtm/pull/906)
+  - `Parameters.SIMPLIFICATION_EXCEPTIONS`
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.18.0)